refactor: consolidate activity-case + memory-ingest into single case job
Both flows share the same downstream (runCASEPipeline + Watch Rules + butler),
so they don't need separate queues / workers / Trigger.dev tasks. Now:

- One queue: "case-queue" (BullMQ + Trigger.dev)
- One worker / task: processCase / caseTask
- One enqueue helper: enqueueCase
- One discriminated payload: { type: "activity" | "memory_ingest", ... }

The two branches are inline `if` cases inside processCase. Adding a third
trigger source = add a branch, no new infrastructure.

Removed: activity-case.{logic.ts,ts}, memory-ingest-case.{logic.ts,ts},
activityCaseQueue, activityCaseWorker, memoryIngestCaseQueue,
memoryIngestCaseWorker, enqueueActivityCase, enqueueMemoryIngestCase.
